Separate env into one that takes only one video...
Separate Binary Classification environment into one base environment that
only works on one single video and doesn't go to the next video when it
resets, and one which works the same way as before (goes to the next
video selected from a generator when it resets).

The reason for this separation is that it becomes easier to re-use and
test the environment when we have full control over which video is used.
Previously, it was impossible to know which video would be loaded next
when the environment reset.

#10